Сообщения к коммитам как вид искусства
